fix(buildkit): compile for 32-bit ARM (Statfs_t.Bsize is int32 there)
Closes opencharly/sdk#262.

`buildkit/privileged_runner.go` multiplied `int64(st.Bavail)` by `st.Bsize`
directly. `syscall.Statfs_t.Bsize` is int64 on 64-bit targets but int32 on
32-bit ones (linux/arm, linux/386), so the mixed-type multiplication only fails
off the 64-bit targets:

    GOOS=linux GOARCH=arm CGO_ENABLED=0 go build ./...
    .../buildkit/privileged_runner.go:75:10: invalid operation:
        int64(st.Bavail) * st.Bsize (mismatched types int64 and int32)

That single line blocked the whole module — and therefore charly and every
plugin importing sdk/buildkit — from building for any 32-bit target.

Fix: convert both operands. The product already promotes to int64, so this is
behaviour-preserving on 64-bit and is exactly what 32-bit needs. The cast is
annotated `//nolint:unconvert` with the reason, because unconvert correctly sees
it as a no-op on amd64 and cannot see the 32-bit target where it is required.

Verified cross-compiling buildkit for all four architectures
(arm, 386, amd64, arm64) and by `go test ./...` + `golangci-lint run ./buildkit`.

TestFreeSpaceArithmeticIs32BitSafe pins the invariant from the source side: a
test cannot observe its own package's compile failure, so the guard asserts the
free-space product converts both operands, catching a refactor that reintroduces
the mixed-type form.

Surfaced by the RDD spike for opencharly/plugin-jetkvm (a static armv7 charly
built from a scratch patch was proven to run on a Rockchip RV1106 JetKVM, which
is what identified this as the sole blocker).

+++ buildkit/privileged_runner.go
@@ -72,7 +72,12 @@ func checkFreeSpace(path string, minBytes int64) error {
 	if err := syscall.Statfs(path, &st); err != nil {
 		return fmt.Errorf("checking free space on %s: %w", path, err)
 	}
-	free := int64(st.Bavail) * st.Bsize
+	// Convert BOTH operands: syscall.Statfs_t.Bsize is int64 on 64-bit targets
+	// but int32 on 32-bit ones (arm, 386), so the mixed-type multiplication
+	// fails to compile for GOOS=linux GOARCH=arm — see opencharly/sdk#262.
+	//
+	//nolint:unconvert // st.Bsize is int32 on 32-bit targets; the cast is a no-op on 64-bit ones and REQUIRED on 32-bit
+	free := int64(st.Bavail) * int64(st.Bsize)
 	if free < minBytes {
 		return fmt.Errorf("insufficient free space on %s: %.1f GiB free, %.1f GiB required (%.0f bytes free, %d bytes required); free disk space and retry",
 			path, float64(free)/gib, float64(minBytes)/gib, float64(free), minBytes)
